# openapi.json
{
  "components": {
    "schemas": {
...
      "GameDetail": {
        "description": "One game, with the two things only a page about *it* has room for.\n\n`season` and `week` aren't schedule facts -- they're where endgame filed\nthis game's play-by-play, and the season file is the only place that says\n(see `app.games.ScheduledGame`). They're on the wire because a reader\nlooking at one game may well want to know which week of which season it\nwas, and because they are what the curve below is fetched by.\n\n`has_win_probability` says whether a fit exists *for the league*, which is\nthe question the page has before it asks for a curve: only football has\none, and asking for a basketball game's would be a request that can only\n404. It does not promise a curve -- the game may have no play-by-play, and\nmost of an NCAAFB week doesn't.",
        "properties": {
          "away": {
            "title": "Away",
            "type": "string"
          },
          "away_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Score"
          },
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "game_id": {
            "title": "Game Id",
            "type": "string"
          },
          "has_win_probability": {
            "title": "Has Win Probability",
            "type": "boolean"
          },
          "home": {
            "title": "Home",
            "type": "string"
          },
          "home_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Score"
          },
          "league": {
            "title": "League",
            "type": "string"
          },
          "market_spread": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Spread"
          },
          "matchup": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchupFacts"
              },
              {
                "type": "null"
              }
            ]
          },
          "neutral": {
            "title": "Neutral",
            "type": "boolean"
          },
          "prediction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GamePrediction"
              },
              {
                "type": "null"
              }
            ]
          },
          "season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week"
          }
        },
        "required": [
          "league",
          "game_id",
          "start",
          "day",
          "home",
          "away",
          "neutral",
          "completed",
          "status",
          "home_score",
          "away_score",
          "market_spread",
          "prediction",
          "season",
          "week",
          "has_win_probability",
          "matchup"
        ],
        "title": "GameDetail",
        "type": "object"
      },
      "GamePrediction": {
        "description": "What one league's default release says about one game.",
        "properties": {
          "away_rating": {
            "title": "Away Rating",
            "type": "number"
          },
          "home_rating": {
            "title": "Home Rating",
            "type": "number"
          },
          "home_win_prob": {
            "title": "Home Win Prob",
            "type": "number"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "predicted_spread": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicted Spread"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "model",
          "run_id",
          "home_win_prob",
          "predicted_spread",
          "home_rating",
          "away_rating"
        ],
        "title": "GamePrediction",
        "type": "object"
      },
      "GameRow": {
        "properties": {
          "away": {
            "title": "Away",
            "type": "string"
          },
          "away_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Score"
          },
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "game_id": {
            "title": "Game Id",
            "type": "string"
          },
          "home": {
            "title": "Home",
            "type": "string"
          },
          "home_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Score"
          },
          "league": {
            "title": "League",
            "type": "string"
          },
          "market_spread": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Spread"
          },
          "neutral": {
            "title": "Neutral",
            "type": "boolean"
          },
          "prediction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GamePrediction"
              },
              {
                "type": "null"
              }
            ]
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "league",
          "game_id",
          "start",
          "day",
          "home",
          "away",
          "neutral",
          "completed",
          "status",
          "home_score",
          "away_score",
          "market_spread",
          "prediction"
        ],
        "title": "GameRow",
        "type": "object"
      },
      "GamesResponse": {
        "properties": {
          "days_ahead": {
            "title": "Days Ahead",
            "type": "integer"
          },
          "days_back": {
            "title": "Days Back",
            "type": "integer"
          },
          "games": {
            "items": {
              "$ref": "#/components/schemas/GameRow"
            },
            "title": "Games",
            "type": "array"
          },
          "since": {
            "format": "date",
            "title": "Since",
            "type": "string"
          },
          "until": {
            "format": "date",
            "title": "Until",
            "type": "string"
          }
        },
        "required": [
          "days_back",
          "days_ahead",
          "since",
          "until",
          "games"
        ],
        "title": "GamesResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryResponse": {
        "properties": {
          "league": {
            "title": "League",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/TeamSeries"
            },
            "title": "Series",
            "type": "array"
          }
        },
        "required": [
          "league",
          "model",
          "run_id",
          "series"
        ],
        "title": "HistoryResponse",
        "type": "object"
      },
...
      "MatchupFacts": {
        "description": "What the matchup terms say about one game.\n\nSigned per side rather than as cassandra's single differential, because a\npage renders two teams and \"the home team is missing its quarterback\" is\nthe fact a reader has; that both being out is worth nothing to either side\nis the model's business, not the page's.\n\n`rest_home` and `rest_away` are None when nobody can say, which is a\nnarrow case and not the same as False. Both are False for two sides on\nlevel rest -- an ordinary week, which is most games. Both are None for a\ngame either side is playing first this season: `RestLedger` prices that at\n0 because \"we have no idea\" and \"level\" are the same number to a model\nabout to add nothing either way, but a page that rendered a season opener\nas \"nobody was rested\" would be stating a fact it doesn't have. Also None\nfor a game whose season the source has no schedule for.",
        "properties": {
          "qb_out_away": {
            "title": "Qb Out Away",
            "type": "boolean"
          },
          "qb_out_home": {
            "title": "Qb Out Home",
            "type": "boolean"
          },
          "rest_away": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rest Away"
          },
          "rest_home": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rest Home"
          }
        },
        "required": [
          "qb_out_home",
          "qb_out_away",
          "rest_home",
          "rest_away"
        ],
        "title": "MatchupFacts",
        "type": "object"
      },
...
      "TeamGamesResponse": {
        "properties": {
          "games": {
            "items": {
              "$ref": "#/components/schemas/TeamGameRow"
            },
            "title": "Games",
            "type": "array"
          },
          "league": {
            "title": "League",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "team": {
            "title": "Team",
            "type": "string"
          }
        },
        "required": [
          "league",
          "model",
          "run_id",
          "team",
          "games"
        ],
        "title": "TeamGamesResponse",
        "type": "object"
      },
...
      "TeamSeries": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/HistoryPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "team": {
            "title": "Team",
            "type": "string"
          }
        },
        "required": [
          "team",
          "points"
        ],
        "title": "TeamSeries",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "invisible-string",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/games": {
      "get": {
        "operationId": "get_games_api_games_get",
        "parameters": [
          {
            "description": "Days of finished games to include, counted in US Central time.",
            "in": "query",
            "name": "back",
            "required": false,
            "schema": {
              "default": 2,
              "description": "Days of finished games to include, counted in US Central time.",
              "maximum": 7,
              "minimum": 0,
              "title": "Back",
              "type": "integer"
            }
          },
          {
            "description": "Days of upcoming games beyond today. 0 is the rest of today.",
            "in": "query",
            "name": "ahead",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Days of upcoming games beyond today. 0 is the rest of today.",
              "maximum": 7,
              "minimum": 0,
              "title": "Ahead",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Games"
      }
    },
    "/api/games/{league}/{game_id}": {
      "get": {
        "description": "Everything this app knows about one game.\n\nThe same row `/api/games` builds, from the same window and the same\nrelease -- a game page that disagreed with the table it was reached from\nwould be worse than no game page. What it adds is the partition its plays\nlive in, and whether there's a model that could draw them.\n\nA game outside the week either side of today needs its `season`, and 404s\nwithout one (see `app.games.find_game`). The horizon still bounds what the\n*window* builds; naming the season buys one game out of one season file\nrather than a wider window, which is what lets a team page link to every\ngame it lists instead of only the fortnight of them in reach.\n\nThe four matchup flags make this a what-if: the same release, asked about\nthe same fixture with a quarterback ruled out or a side off a bye. They are\nrefused on a game that has been played, which is not squeamishness about\ncounterfactuals -- a completed game shows the forecast that was made before\nit (see the module docstring), and that forecast is a stored number no flag\ncan reach. Answering a what-if with it would be silently ignoring the\nrequest.",
        "operationId": "get_game_api_games__league___game_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "league",
            "required": true,
            "schema": {
              "title": "League",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "description": "The season the game is in. Only needed for a game outside the week either side of today, which can't be found without it.",
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The season the game is in. Only needed for a game outside the week either side of today, which can't be found without it.",
              "title": "Season"
            }
          },
          {
            "description": "Home QB out. Only for a game that hasn't been played, in a league with the term.",
            "in": "query",
            "name": "qb_out_home",
            "required": false,
            "schema": {
              "default": false,
              "description": "Home QB out. Only for a game that hasn't been played, in a league with the term.",
              "title": "Qb Out Home",
              "type": "boolean"
            }
          },
          {
            "description": "Away QB out. Only for a game that hasn't been played, in a league with the term.",
            "in": "query",
            "name": "qb_out_away",
            "required": false,
            "schema": {
              "default": false,
              "description": "Away QB out. Only for a game that hasn't been played, in a league with the term.",
              "title": "Qb Out Away",
              "type": "boolean"
            }
          },
          {
            "description": "Home side off the longer break. Only for a game that hasn't been played, in a league with the term.",
            "in": "query",
            "name": "rest_home",
            "required": false,
            "schema": {
              "default": false,
              "description": "Home side off the longer break. Only for a game that hasn't been played, in a league with the term.",
              "title": "Rest Home",
              "type": "boolean"
            }
          },
          {
            "description": "Away side off the longer break. Only for a game that hasn't been played, in a league with the term.",
            "in": "query",
            "name": "rest_away",
            "required": false,
            "schema": {
              "default": false,
              "description": "Away side off the longer break. Only for a game that hasn't been played, in a league with the term.",
              "title": "Rest Away",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Game"
      }
    },
...
    "/api/leagues/{league}/history": {
      "get": {
        "operationId": "get_history_api_leagues__league__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "league",
            "required": true,
            "schema": {
              "title": "League",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated team names, at most 5. Named exactly as the ratings table spells them.",
            "in": "query",
            "name": "teams",
            "required": true,
            "schema": {
              "description": "Comma-separated team names, at most 5. Named exactly as the ratings table spells them.",
              "title": "Teams",
              "type": "string"
            }
          },
          {
            "description": "Defaults to the league's lowest-Brier model.",
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Defaults to the league's lowest-Brier model.",
              "title": "Model"
            }
          },
          {
            "description": "First season year to include. Defaults to all of them.",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First season year to include. Defaults to all of them.",
              "title": "From"
            }
          },
          {
            "description": "Last season year to include. Defaults to all of them.",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last season year to include. Defaults to all of them.",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get History"
      }
    },
...
    "/api/leagues/{league}/teams/{team}/games": {
      "get": {
        "description": "Every game this model has a stored prediction for, for one team.\n\nThe whole career in one response, for the reason the history endpoint\nserves a whole line: it is a few hundred rows, which is smaller than the\nratings table the page was reached from, and a season picker that went\nback to the network to narrow what it already had would be slower and no\nsmaller.\n\nThe model follows the page's selection, like the chart above it. A game\nlist showing elo's forecasts under glicko's rank would be two models\nwearing one heading.",
        "operationId": "get_team_games_api_leagues__league__teams__team__games_get",
        "parameters": [
          {
            "in": "path",
            "name": "league",
            "required": true,
            "schema": {
              "title": "League",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "team",
            "required": true,
            "schema": {
              "title": "Team",
              "type": "string"
            }
          },
          {
            "description": "Defaults to the league's lowest-Brier model.",
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Defaults to the league's lowest-Brier model.",
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamGamesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Team Games"
      }
    },
...
  }
}